packages/requirements.txt
python-telegram-bot==21.10
openai==1.63.2
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.2
nest-asyncio==1.6.0
python-telegram-bot[webhooks]